+++ src/Oscar.jl
@@ -228,6 +228,9 @@ include("TropicalGeometry/TropicalGeometry.jl")
 
 include("InvariantTheory/InvariantTheory.jl")
 
+# Serialization should always come at the end of Oscar source code
+# but before experimental, any experimental serialization should
+# be written inside the corresponding experimental code sub directory
 include("Serialization/main.jl")
 
 include("../experimental/Experimental.jl")
